requirements.txt
Django==2.0.3
django-cors-headers==2.2.0
django-filter==1.1.0
djangorestframework==3.7.7
pytz==2018.3